Cleanup unused functions/imports after refactor
