+++ blocks/antelope/src/lib.rs
@@ -7,11 +7,10 @@ mod blocks;
 mod creation_tree;
 mod db_ops;
 mod feature_ops;
-mod keys;
+mod map_events;
 mod pb;
 mod perm_ops;
 mod ram_ops;
-mod sinks;
 mod size;
 mod table_ops;
 mod transactions;
