+++ blocks/antelope/src/perm_ops.rs
@@ -1,16 +1,6 @@
-use crate::{
-    authority::insert_authority,
-    keys::perm_ops_keys,
-    pb::antelope::PermOp as RawPermOp,
-    transactions::{insert_transaction_metadata, is_transaction_success},
-};
-use common::{blocks::insert_timestamp, structs::BlockTimestamp};
-use substreams::pb::substreams::Clock;
-use substreams_antelope::{
-    pb::{PermOp, TransactionTrace},
-    Block,
-};
-use substreams_database_change::pb::database::{table_change, DatabaseChanges};
+use crate::{pb::antelope::PermOp as RawPermOp, transactions::is_transaction_success};
+use common::structs::BlockTimestamp;
+use substreams_antelope::Block;
 
 pub fn perm_op_operation_to_string(operation: i32) -> String {
     match operation {
@@ -19,43 +9,6 @@ pub fn perm_op_operation_to_string(operation: i32) -> String {
         2 => "Update".to_string(),
         3 => "Remove".to_string(),
         _ => "Unknown".to_string(),
-    }
-}
-
-pub fn insert_perm_op(tables: &mut DatabaseChanges, clock: &Clock, transaction: &TransactionTrace, perm_op: &PermOp) {
-    // transaction
-    let tx_hash = &transaction.id;
-
-    // action
-    let action_index = perm_op.action_index;
-
-    // perm_op
-    let operation = perm_op_operation_to_string(perm_op.operation);
-    match &perm_op.new_perm {
-        Some(new_perm) => {
-            let keys = perm_ops_keys(tx_hash, &action_index);
-            let threshold = new_perm.authority.as_ref().map_or(0, |authority| authority.threshold);
-            let row = tables
-                .push_change_composite("perm_ops", keys, 0, table_change::Operation::Create)
-                .change("operation", ("", operation.as_str()))
-                .change("operation_code", ("", perm_op.operation.to_string().as_str()))
-                .change("action_index", ("", action_index.to_string().as_str()))
-                .change("id", ("", new_perm.id.to_string().as_str()))
-                .change("parent_id", ("", new_perm.parent_id.to_string().as_str()))
-                .change("owner", ("", new_perm.owner.to_string().as_str()))
-                .change("name", ("", new_perm.name.to_string().as_str()))
-                .change("threshold", ("", threshold.to_string().as_str()));
-            insert_transaction_metadata(row, transaction);
-            insert_timestamp(row, clock, false, false);
-
-            match &new_perm.authority {
-                Some(authority) => {
-                    insert_authority(tables, clock, transaction, action_index, authority);
-                }
-                None => {}
-            }
-        }
-        None => {}
     }
 }
