+++ blocks/antelope/src/table_ops.rs
@@ -1,16 +1,7 @@
-use common::{blocks::insert_timestamp, structs::BlockTimestamp};
-use substreams::pb::substreams::Clock;
-use substreams_antelope::{
-    pb::{TableOp, TransactionTrace},
-    Block,
-};
-use substreams_database_change::pb::database::{table_change, DatabaseChanges};
+use common::structs::BlockTimestamp;
+use substreams_antelope::Block;
 
-use crate::{
-    keys::table_ops_keys,
-    pb::antelope::TableOp as RawTableOp,
-    transactions::{insert_transaction_metadata, is_transaction_success},
-};
+use crate::{pb::antelope::TableOp as RawTableOp, transactions::is_transaction_success};
 
 pub fn table_op_operation_to_string(operation: i32) -> String {
     match operation {
@@ -19,33 +10,6 @@ pub fn table_op_operation_to_string(operation: i32) -> String {
         2 => "Remove".to_string(),
         _ => "Unknown".to_string(),
     }
-}
-
-pub fn insert_table_op(tables: &mut DatabaseChanges, clock: &Clock, transaction: &TransactionTrace, table_op: &TableOp, index: &u32) {
-    let action_index = &table_op.action_index;
-    let operation = &table_op_operation_to_string(table_op.operation);
-    let payer = &table_op.payer;
-    let code = &table_op.code;
-    let scope = &table_op.scope;
-    let table_name = &table_op.table_name;
-
-    // transaction
-    let tx_hash = &transaction.id;
-
-    let keys = table_ops_keys(tx_hash, index);
-    let row = tables
-        .push_change_composite("table_ops", keys, 0, table_change::Operation::Create)
-        .change("index", ("", index.to_string().as_str()))
-        .change("action_index", ("", action_index.to_string().as_str()))
-        .change("operation", ("", operation.as_str()))
-        .change("operation_code", ("", table_op.operation.to_string().as_str()))
-        .change("payer", ("", payer.as_str()))
-        .change("code", ("", code.as_str()))
-        .change("scope", ("", scope.as_str()))
-        .change("table_name", ("", table_name.as_str()));
-
-    insert_transaction_metadata(row, transaction);
-    insert_timestamp(row, clock, false, false);
 }
 
 pub fn collect_table_ops(block: &Block, timestamp: &BlockTimestamp) -> Vec<RawTableOp> {
